requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.5.0
pymongo==4.8.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
pydantic-settings==2.4.0
jinja2==3.1.4
python-multipart==0.0.9
beautifulsoup4==4.12.3
passlib==1.7.4
